+++ index.ts
@@ -2,17 +2,140 @@ import {
   getArticleById,
   getArticleByPath,
   getArticles,
-  getLatestArticles 
+  getLatestArticles, 
+  getVideoArticles
 } from './src/functions/articles';
 
-import { getUserById } from './src/functions/user';
+import {
+  getArticleComments,
+  getCommentById
+} from './src/functions/comments';
+
+import {
+  getListings
+} from './src/functions/listing'
+
+import { 
+  getUserById 
+} from './src/functions/users';
+
 import DevToClient from "./src/structures/DevToClient";
+
+import {
+  AuthorizationOptions,
+  BaseArticle,
+  BaseFlareTag,
+  BaseFollower,
+  BaseListing,
+  BaseNewArticleData,
+  BaseOrganization,
+  BasePageFetchOptions,
+  BaseUser,
+  BaseVideoArticle,
+  RawArticle,
+  RawComment,
+  RawFlareTag,
+  RawFollower,
+  RawListing,
+  RawNewArticleData,
+  RawOrganization,
+  RawUser,
+  RawVideoArticle,
+  DevDotToErrorResponse,
+  Comment,
+  ListingCategoryOptions,
+  FinalArticleIdentifierOptions,
+  FinalPageFetchOptions,
+  FinalPodcastIdentifierOptions,
+  FlareTag,
+  FollowedTags,
+  Follower,
+  PageFetchOptions,
+  PodcastIdentifierOptions,
+  PostOptions,
+  Organization,
+  SortOptions,
+  Article,
+  ArticleIdentifierOptions,
+  NewArticleData,
+  VideoArticle
+} from './src/typings/interfaces';
+
+import {
+  ArticleState,
+  TypeOfArticle,
+  EndPoint,
+  NumberResolvable,
+  Page,
+  SortedPageOptions,
+  BaseFetchPageOptions,
+  ListingCategory,
+  CategorizedListingOptions,
+  Tags,
+  TypeOfMember,
+  StringIndex,
+  VideoDurationFormat
+} from './src/typings/types';
 
 export {
   getArticles,
   getLatestArticles,
   getArticleByPath,
   getArticleById,
+  getVideoArticles,
+  getListings,
   getUserById,
-  DevToClient
+  getArticleComments,
+  getCommentById,
+  DevToClient,
+  AuthorizationOptions,
+  BaseArticle,
+  BaseFlareTag,
+  BaseFollower,
+  BaseListing,
+  BaseNewArticleData,
+  BaseOrganization,
+  BasePageFetchOptions,
+  BaseUser,
+  BaseVideoArticle,
+  RawArticle,
+  RawComment,
+  RawFlareTag,
+  RawFollower,
+  RawListing,
+  RawNewArticleData,
+  RawOrganization,
+  RawUser,
+  RawVideoArticle,
+  DevDotToErrorResponse,
+  Comment,
+  ListingCategoryOptions,
+  FinalArticleIdentifierOptions,
+  FinalPageFetchOptions,
+  FinalPodcastIdentifierOptions,
+  FlareTag,
+  FollowedTags,
+  Follower,
+  PageFetchOptions,
+  PodcastIdentifierOptions,
+  PostOptions,
+  Organization,
+  SortOptions,
+  Article,
+  ArticleIdentifierOptions,
+  NewArticleData,
+  VideoArticle,
+  ArticleState,
+  TypeOfArticle,
+  EndPoint,
+  NumberResolvable,
+  Page,
+  SortedPageOptions,
+  BaseFetchPageOptions,
+  ListingCategory,
+  CategorizedListingOptions,
+  Tags,
+  TypeOfMember,
+  StringIndex,
+  VideoDurationFormat
 };