+++ src/functions/articles.ts
@@ -1,8 +1,19 @@
-import { Article, PageFetchOptions } from "../typings/interfaces";
+import { Article, PageFetchOptions, VideoArticle } from "../typings/interfaces";
 import { BaseFetchPageOptions, NumberResolvable, Page } from "../typings/types";
 import DevDotToUtil from "../structures/DevDotToUtil";
 
 const { request, parseParameters } = DevDotToUtil;
+
+ /**
+ * Fetches a page of articles.
+ * @param {Partial<PageFetchOptions>} options Describes the target articles.
+ * @returns {Promise<Page>} An array of articles.
+ */
+  export async function 
+  getArticles(options?:Partial<PageFetchOptions>):Promise<Page<Article>> {
+    const query = parseParameters(options);
+    return await request(`/articles${query}`);
+  }
 
 /**
  * Fetches an article by it's id.
@@ -11,12 +22,7 @@ const { request, parseParameters } = DevDotToUtil;
  */
 export async function 
 getArticleById(id:NumberResolvable):Promise<Article> {
-  try {
-    const req = await request<Article>(`/articles/${id}`);
-    return req;
-  } catch(err) {
-    throw err;
-  }
+  return await request(`/articles/${id}`);
 }
 
 /**
@@ -27,45 +33,29 @@ getArticleById(id:NumberResolvable):Promise<Article> {
  */
 export async function 
 getArticleByPath(publisherName: string, slug:string):Promise<Article> {
-  try {
-    const req = await request<Article>(`/articles/${publisherName}/${slug}`);
-    return req;
-  } catch(err) {
-    throw err;
-  }
-}
-
- /**
- * Fetches a page of articles.
- * @param {Partial<PageFetchOptions>} options Describes the target articles.
- * @returns {Promise<Page>} An array of articles.
- */
-export async function 
-getArticles(options?:Partial<PageFetchOptions>):Promise<Page> {
-  try {
-    const query = parseParameters(options);
-    const req = request<Page>(`/articles${query}`);
-    
-    return req;
-  } catch(err) {
-    throw err;
-  }
+  return await request(`/articles/${publisherName}/${slug}`);
 }
 
 /**
  * Fetches the newest articles.
  * @param {BaseFetchPageOptions} options 
- * @returns {Promise<Page>} An array of articles
+ * @returns {Promise<Page<Article>>} An array of articles
  */
 export async function getLatestArticles(
   options?: BaseFetchPageOptions
-):Promise<Page> {
-  try {
-    const query = parseParameters(options);
-    const req = request<Page>(`/articles/latest${query}`);
-    
-    return req; 
-  } catch(err) {
-    throw err;
-  }
+):Promise<Page<Article>> {
+  const query = parseParameters(options);
+  return request(`/articles/latest${query}`);
+}
+
+/**
+ * Fetches video articles.
+ * @param {BaseFetchPageOptions} options 
+ * @returns {Promise<Page<VideoArticle>>} An array of articles
+ */
+export async function getVideoArticles(
+  options?: BaseFetchPageOptions
+):Promise<Page<VideoArticle>> {
+  const query = parseParameters(options);
+  return request(`/videos${query}`);
 }