+++ src/typings/types.ts
@@ -1,11 +1,17 @@
-import { Article, PageFetchOptions } from "./interfaces";
+import { ListingCategoryOptions, PageFetchOptions, SortOptions } from "./interfaces";
 
 export type ArticleState = 'all' | 'fresh' | 'rising';
+export type CategorizedListingOptions = BaseFetchPageOptions & ListingCategoryOptions;
 export type EndPoint = `/${string}`;
 export type NumberResolvable = number | `${number}`;
-export type Page = Article[];
+export type Page<T> = T[];
 export type BaseFetchPageOptions = Partial<Pick<PageFetchOptions, 'perPage' | 'page'>>;
+export type ListingCategory = 'cfp' | 'forhire' | 'collabs' | 'education' | 'jobs' | 'mentors' | 'products' | 'mentees' | 'forsale' | 'events' | 'misc';
 export type Tags = string;
-export type TypeOfArticle = 'article';
+export type TypeOfArticle = 'article' | 'video_article';
 export type TypeOfMember = 'user';
-export type StringIndex<T> = { [key:string]: string }
+export type StringIndex<T> = { [key:string]: string };
+export type SortedPageOptions = BaseFetchPageOptions & SortOptions;
+export type VideoDurationFormat = 
+  `${number}${number}:${number}${number}` | 
+  `${number}${number}:${number}${number}:${number}${number}`;